+++ src/pages/Home/partials/CarsCard/index.styles.ts
